GENIO: board: mt8390: add board id judging to support multi panel
On the G700 EVK p1v4, a different panel, KD070FHFI-D078 is utilized.
To ensure compatibility of the G700 EVK image with both panels, this
modification checks the board ID and selects the appropriate DSI driver
and dtbo.

The added functions are described below.

check_board_id():
The board ID can be verified by measuring the voltage input of AUX IN
channel 1. If the voltage exceeds 1100 mV, it indicates that this board
is a p1v4.

panel_get_desc():
This function provides the DSI panel driver for bootlogo and was
previously defined in mt8390_evk_panel.c. As there are two panel
drivers, the panel_get_desc() in mt8390_evk_panel.c is modified to
panel_get_desc_kd070fhfid015, and a new panel_get_desc() is defined in
mt8390_evk.c. The new panel_get_desc() invokes check_board_id() to
determine whether to use kd070fhfid078 or kd070fhfid015.

board_late_init():
This function is the final function called before entering the Linux
kernel. It invokes check_board_id() to determine whether to load the
display-dsi-p1v4.dtbo in order to enable kd070fhfid078 to function
within the Linux kernel.

Add a mt8390_evk.h to define specific elements for clarity, including
the ADC node name, the AUX channel to be read, the AUX threshold used
to determine the board ID, and the board ID.

Change-Id: I3884e9b023b0fc4d28ebb2011edf1f0b4aaab1c9
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>